api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dateutil
psycopg2-binary
pandas
pyyaml
celery[redis]
aiofiles
